+++ templates/base.html
@@ -106,16 +106,21 @@
   <div class="container mt-5">
     <div class="row">
       <div class="col-lg">
-      {% with messages = get_flashed_messages(with_categories=true) %}
-      {% if messages %}
-      {% for category, message in messages %}
-      <div class="alert alert-{{ category }} alert-dismissible fade show" role="alert">
-         <strong>{{ message }}</strong>
-         <button type="button" class="btn-close" data-dismiss="alert" aria-label="Close"></button>
-      </div>
-      {% endfor %}
-      {% endif %}
-      {% endwith %}
+        {%if active_page != 'login' or active_page != 'register'%}
+        {% with messages = get_flashed_messages(with_categories=true) %}
+        {% if messages %}
+        {% for category, message in messages %}
+        <div class="alert alert-{{ category }} alert-dismissible fade show" role="alert">
+           <strong>{{ message }}</strong>
+           <button type="button" class="btn-close" data-dismiss="alert" aria-label="Close"></button>
+        </div>
+        {% endfor %}
+        {% endif %}
+        {% endwith %}
+        {%else%}
+        {%block FLASH%}
+        {%endblock%}
+      {%endif%}
     </div>
   </div>
 </div>
