+++ templates/events.html
@@ -2,24 +2,61 @@
 {%block title%} EVENTS {%endblock%}
 {%set active_page='events'%}
 {%block content%} 
+<div class="container" style="margin-top: 80px; margin-left:30px;margin-right:30px;justify-content:center">
+    <div class="event-cards">
 
-<div class="event-cards">
+
+    <div class="row row-events">
     
-    <div class="card">
+    <div class="col-6">
+        <div class="card card-event">
+            <img class="event-img" src="static/assets/img/wednesday!.jpg"> 
+            <h1 class="event-title"> Project Wednesdays</h1>
+            <p class="event-desc"> 
+                Project Wednesdays are a perfect opportunity for students to spend time exploring other interests outside of class. Sign up for your Project Wednesday here!
+            </p>
+        </div>
+    </div>
+
+    <div class="col-6">
+        <div class="card card-event">
+            <img class="event-img" src="static/assets/img/wednesday!.jpg">
+            <h1 class="event-title"> Cultural Day Workshops </h1>
+            <p class="event-desc"> 
+                Lorem ipsum dolor sit amet consectetur adipisicing elit. Sapiente harum debitis tenetur sed ipsum ducimus error delectus, minima consectetur reiciendis repudiandae itaque, praesentium dicta eius doloremque aspernatur facere iusto eveniet.
+            </p>
+        </div>
+    </div>
 
     </div>
 
-    <div class="card">
+
+    <div class="row row-events">
+
+        <div class="col-6">
+            <div class="card card-event">
+                <img class="event-img" src="static/assets/img/wednesday!.jpg">
+                <h1 class="event-title"> Coffeehouse</h1>
+                <p class="event-desc"> 
+                    Lorem ipsum dolor sit, amet consectetur adipisicing elit. Quas incidunt, at dolores natus error itaque? Assumenda nemo adipisci error necessitatibus porro. Laborum suscipit nostrum consequatur ipsum, sapiente nisi aut dignissimos, iste tempore nesciunt cum quos fugit vero culpa illum quidem inventore omnis quae! Temporibus repudiandae voluptatem dolor laboriosam, alias corporis?
+                </p>
+            </div>
+        </div>
+
+        <div class="col-6">
+            <div class="card card-event">
+                <img class="event-img" src="static/assets/img/wednesday!.jpg">
+                <h1 class="event-title"> BSU Volleyball Tournament</h1>
+                <p class="event-desc"> 
+                    You should participate :>
+                </p>
+            </div>
+        </div>
 
     </div>
 
-    <div class="card">
 
-    </div>
-
-    <div class="card">
-
-    </div>
 </div>
-
+</div>
 {%endblock%}
+{%block main_css%} <link href="static/assets/css/style.css" rel="stylesheet"> {%endblock%}
