+++ templates/sampledate.html
@@ -1,67 +1,8 @@
 {%extends "base.html"%}
-{%block title %} Sample Date Page {%endblock%}
+{%block title %} Events Page {%endblock%}
 {%block main_css%} <link href="static/assets/css/style2.css" rel="stylesheet"> {%endblock%}
 {%set active_page = 'calendar'%}
 {%block content %}
 
-
-<link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.3.0/css/font-awesome.min.css" rel="stylesheet"> 
-<div class="container">
-<div class="row">
-  <div class="col-md-10">
-    <div class="box box-aqua">
-      <div class="box-header ui-sortable-handle" style="cursor: move;">
-        <i class="ion ion-clipboard"></i>
-        <h3 class="box-title"> November 1st, 2017</h3>
-        
-      </div>
- 
-      <div class="box-body">
-        <ul class="todo-list ui-sortable">
-          <li>
-            <span class="handle ui-sortable-handle">
-              <i class="fa fa-ellipsis-v"></i>
-              <i class="fa fa-ellipsis-v"></i>
-            </span>
-            <input type="checkbox" value="" name="">
-            <span class="text">Event 1</span>
-            <div class="tools">
-              <i class="fa fa-edit"></i>
-              <i class="fa fa-trash-o"></i>
-            </div>
-          </li>
-          <li>
-            <span class="handle ui-sortable-handle">
-              <i class="fa fa-ellipsis-v"></i>
-              <i class="fa fa-ellipsis-v"></i>
-            </span>
-            <input type="checkbox" value="" name="">
-            <span class="text"> <a href = '/edit_event'> Event 2 </a> </span>
-              <div class="tools">
-              <a href = '/edit_event'> <i class="fa fa-edit"> </i> </a>
-              <i class="fa fa-trash-o"></i>
-            </div>
-          </li>
-
-          <li>
-            <span class="handle ui-sortable-handle">
-              <i class="fa fa-ellipsis-v"></i>
-              <i class="fa fa-ellipsis-v"></i>
-            </span>
-            <input type="checkbox" value="" name="">
-            <span class="text">Event 3</span>
-            <div class="tools">
-              <i class="fa fa-edit"></i>
-              <i class="fa fa-trash-o"></i>
-            </div>
-          </li>
-        </ul>
-      </div>
-    <!-- <div class="box-footer clearfix no-border">-->
-    <!--  <button type="button" class="btn btn-default pull-right"><i class="fa fa-plus"></i> Add item</button> -->
-    <!--  </div> -->
-    </div>
-  </div> 
-  
  
 {%endblock%}
